Split the SierraSubjects trait
