+++ catalogue_pipeline/transformer/src/test/scala/uk/ac/wellcome/platform/transformer/utils/SierraDataUtil.scala
@@ -28,6 +28,20 @@ trait SierraDataUtil extends IdentifiersUtil with SierraUtil {
 
   def createSierraBibData: SierraBibData = createSierraBibDataWith()
 
+  def bibData(marcTag: String, marcSubfields: List[MarcSubfield]) = {
+    createSierraBibDataWith(
+      varFields = List(
+        VarField(
+          fieldTag = "p",
+          marcTag = marcTag,
+          indicator1 = "",
+          indicator2 = "",
+          subfields = marcSubfields
+        )
+      )
+    )
+  }
+
   def createSierraItemDataWith(
     deleted: Boolean = false,
     location: Option[SierraLocation] = None
